+++ js/checking-length.js
@@ -1,0 +1,6 @@
+// Функция для проверки длины строки.
+const checkingLength = function (str, maxLength){
+  return str.length <= maxLength;
+};
+
+export{checkingLength};
